refactor(asr): extract calibration into _calibration.py
Move calibrate_asr and _fit_component_thresholds to _calibration.py, importing
the window/distribution/covariance/filter/SPD helpers from their new homes.
core.py re-exports via __all__; no behavior change.
